Install the command line tool via setuptools' console_scripts entrypoint
The previous option of installing it via scripts= didn't add a shebang to the
file, meaning that if you actually tried to call it directly, it was interpreted
as a shell script. Using the console_scripts entrypoint to install it
automatically installs a wrapper scripts that calls the main function that has
the correct shebang (for example in a virtualenv).