UpnpEventHandler only track services, not timeouts
UpnpEventHandler does not use the timeout or renewal_time information in
its internal _subscriptions dict, so simplify it to a simple sid:service
mapping.

Use a WeakValueDictionary to prevent keeping UpnpServices alive when the
owning UpnpDevice is deleted.

Also, allow resubscribing or unsubscribing by SID as well as UpnpService,
to reduce need for reverse lookups in the _subscriptions dict.