Automatic service resubscriptions (#73)
* Only unsubscribe device's own services, not for all devices

* Run event un-/re-subscriptions in parallel

* mypy change: Returned headers are Mapping[str, str]

* Consistent naming of AddressFamily.AF_INET

* UpnpEventHandler only track services, not timeouts

UpnpEventHandler does not use the timeout or renewal_time information in
its internal _subscriptions dict, so simplify it to a simple sid:service
mapping.

Use a WeakValueDictionary to prevent keeping UpnpServices alive when the
owning UpnpDevice is deleted.

Also, allow resubscribing or unsubscribing by SID as well as UpnpService,
to reduce need for reverse lookups in the _subscriptions dict.

* Wrap aiohttp and asyncio exceptions with own

The use of aiohttp is an implementation detail, so don't leak its
exceptions to users of this library. The wrapping exceptions are derived
from the wrapped ones, so should still be catchable by users expecting
aiohttp exceptions.

* UpnpEventHandler's subcription methods raise exceptions for errors

Instead of returning None or False to indicate an error occurred, raise an
exception that gives better detail. This makes it consistent, as exceptions
could already be raised (or passed through) from aiohttp failures.

* UpnpProfileDevice can resubscribe automatically

Resubscriptions are done by an async Task that resubscribes when the
renewal time has almost approached.

Failed resubscriptions are deleted, and the `on_event` callback called to
inform listeners that the device might be offline. In that case, the
services should be resubscribed when the device is available again.

* Allow pytest args through tox command lines

Allows e.g. running one particular test
(`tox -e py36 -- tests/test_ssdp.py`),
or capturing debug logs (`--log-level=DEBUG`).

* Tests for `UpnpProfileDevice.async_subscribe_services`

* UpnpEventHandler.async_unsubscibe always deletes subscriptions

Always delete a subscription from internal mapping, even if unsubscribe
fails. Failure most likely means the device is offline or not supporting
subscriptions, so keeping the SID is pointless.

* UpnpEventHandler resubscribe tries subscribe too

In case a resubscribe call fails, try a plain subscribe, in case the device
has changed location or lost a subscription somehow.

* Update CHANGES.rst

* Report subscription status via is_subscribed

* Minimal polling in DmrDevice.async_update

Add an option to DmrDevice.async_update to only do the bare-minimum polling
required, for when event subscriptions are sending notifications instead.