Retry session requests for ClientConnectionError
The server is allowed to disconnect at any time during a request session,
at which point we want to retry the request.

A disconnection could manifest as an aiohttp.ServerDisconnectedError if
it happened between requests, or it could be aiohttp.ClientOSError if it
happened while we are writing the request body. Both errors derive from
aiohttp.ClientConnectionError for socket errors.