DmrDevice helper method improvements (#142)
* Poll CurrentTransportActions in async_wait_for_can_play

Some devices may not send event notifications for changes to
`CurrentTransportActions`, in which case waiting for the variable to change
will be futile. So try polling (via `GetCurrentTransportActions`) to get
the latest state.

* Mark test_server_init as async for event loop

Fix "DeprecationWarning: There is no current event loop" reported in tests.

* Less network usage in DmrDevice._fetch_headers

If HTTP HEAD fails, try GET with a Range of 0-0 (1 byte), and if that fails
then fallback to GET of the whole file. Also bail out early if response
code of 404 (Not Found) is received, indicating file does not exist.

* towncrier

* Fix pylint and Python 3.8 errors

* Python 3.8 does not support `|` operand for `dict`s.

* Skip test requiring AsyncMock on Python <= 3.7